+++ package.json
@@ -24,6 +24,7 @@
         "@babel/plugin-proposal-optional-chaining": "7.9.0",
         "@babel/preset-env": "7.9.6",
         "@babel/preset-typescript": "7.9.0",
+        "@types/eslint-config-prettier": "^6.11.0",
         "@types/glob": "7.1.1",
         "@types/jest": "25.2.1",
         "@types/node": "12.12.21",
