+++ src/creation/writeConversionResults.ts
@@ -1,6 +1,7 @@
 import { FileSystem } from "../adapters/fileSystem";
 import { AllOriginalConfigurations } from "../input/findOriginalConfigurations";
 import { createEnv } from "./eslint/createEnv";
+import { trimESLintRules } from "./eslint/trimESLintRules";
 import { formatConvertedRules } from "./formatConvertedRules";
 import { formatOutput } from "./formatting/formatOutput";
 import { SummarizedResultsConfiguration } from "./summarization/types";
@@ -34,7 +35,7 @@ export const writeConversionResults = async (
         },
         plugins,
         rules: {
-            ...formatESLintRules(eslint?.full.rules, summarizedResults),
+            ...trimESLintRules(eslint?.full.rules, summarizedResults.extensionRules),
             ...formatConvertedRules(summarizedResults, tslint.full),
         },
     };
