+++ src/rules/formats/convertRuleSeverity.test.ts
@@ -1,39 +1,4 @@
-import { convertTSLintRuleSeverity, convertRawESLintRuleSeverity } from "./convertRuleSeverity";
-
-describe("convertRuleSeverity", () => {
-    it("returns error when the severity is error", () => {
-        // Arrange
-        const tslintSeverity = "error";
-
-        // Act
-        const eslintSeverity = convertTSLintRuleSeverity(tslintSeverity);
-
-        // Assert
-        expect(eslintSeverity).toEqual("error");
-    });
-
-    it("returns off when the severity is off", () => {
-        // Arrange
-        const tslintSeverity = "off";
-
-        // Act
-        const eslintSeverity = convertTSLintRuleSeverity(tslintSeverity);
-
-        // Assert
-        expect(eslintSeverity).toEqual("off");
-    });
-
-    it("returns warn when the severity is warning", () => {
-        // Arrange
-        const tslintSeverity = "warning";
-
-        // Act
-        const eslintSeverity = convertTSLintRuleSeverity(tslintSeverity);
-
-        // Assert
-        expect(eslintSeverity).toEqual("warn");
-    });
-});
+import { convertRawESLintRuleSeverity } from "./convertRuleSeverity";
 
 describe("convertRawESLintRuleSeverity", () => {
     it("returns off when the severity is 0", () => {
