+++ src/rules/formats/convertTSLintRuleSeverity.ts
@@ -1,0 +1,10 @@
+import { ESLintRuleSeverity, TSLintRuleSeverity } from "../types";
+
+/**
+ * Converts a TSLint rule severity string to the ESLint equivalent.
+ */
+export const convertTSLintRuleSeverity = (
+    tslintSeverity: TSLintRuleSeverity,
+): ESLintRuleSeverity => {
+    return tslintSeverity === "warning" ? "warn" : tslintSeverity;
+};
