--- src/typings/eslint-config-prettier.d.ts
@@ -1,4 +1,0 @@
-// Pending https://github.com/DefinitelyTyped/DefinitelyTyped/pull/44245...
-declare module "eslint-config-prettier" {
-    export const rules: Record<string, 0 | "off">;
-}
